api/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
Deprecated==1.3.1
ecdsa==0.19.1
Faker==40.8.0
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
standardwebhooks==1.0.1
starlette==0.52.1
svix==1.86.0
types-Deprecated==1.3.1.20260130
types-python-dateutil==2.9.0.20260305
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
wrapt==2.1.2
